PIM-10985: Mark skipped migrations as executed

+++ upgrades/schema/Version_6_0_20210331105730_remove_datagrid_view_unique_label_constraint.php
@@ -34,7 +34,12 @@ final class Version_6_0_20210331105730_remove_datagrid_view_unique_label_constra
         SQL;
         $uniqueConstraintKeyName = $this->dbalConnection()->executeQuery($selectUniqueConstraint)->fetchOne();
 
-        $this->skipIf(!$uniqueConstraintKeyName, 'pim_datagrid_view unique constraint is already up to date');
+        if (!$uniqueConstraintKeyName) {
+            $this->write('pim_datagrid_view unique constraint is already up to date');
+
+            return;
+        }
+
         $this->addSql("ALTER TABLE pim_datagrid_view DROP index $uniqueConstraintKeyName");
     }
